+++ src/Control/ElementFormController.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace DNADesign\Elemental\UserForms\Control;
+
+use DNADesign\Elemental\Controllers\ElementController;
+use SilverStripe\Control\Controller;
+use SilverStripe\UserForms\Control\UserDefinedFormController;
+
+class ElementFormController extends ElementController
+{
+    public function __construct($element = null)
+    {
+        parent::__construct($element);
+
+        $current = Controller::curr();
+
+        if ($current->getRequest()->isPOST()) {
+            // handle the post request.
+            $user = UserDefinedFormController::create($element);
+            $form = $user->Form();
+
+            $user->process($current->getRequest()->postVars(), $form);
+        }
+    }
+}
